#14 Add draft of receiver configuration
