+++ CMakeLists.txt
@@ -69,7 +69,8 @@ target_include_directories(qctool_lib PUBLIC "${DCMTK_INCLUDE_DIRS}")
 
 enable_testing()
 
-add_executable(unit_tests tests/MyTests.cpp)
+add_executable(unit_tests tests/MyTests.cpp 
+                          tests/ReceiverTests.cpp)
 target_include_directories(unit_tests PRIVATE tests)
 target_link_libraries(unit_tests PRIVATE qctool_lib)
 add_test(run_unit_tests unit_tests)
