+++ src/draft_receiver.hpp
@@ -21,6 +21,11 @@ public:
     /** Overwrite method of DcmSCP to enable handling of C-STORE requests. */
     OFCondition handleIncomingCommand(T_DIMSE_Message* incomingMsg, const DcmPresentationContextInfo& presInfo);
 
+    
+    //void setIPs();
+    /**Check if calling IP is accepted */
+    //virtual OFBool checkCallingHostAccepted();
+
 };
 
 /**
@@ -28,6 +33,7 @@ public:
  */
 class Receiver : public DcmSCPPool<ReceiverThread>, public OFThread
 {
+    OFList<OFString> m_sourcelist;
 
 public:
     OFCondition result;
@@ -41,9 +47,19 @@ public:
     /** Stop listening after current association. */
     virtual void request_stop();
 
+    /** Set AE Title. */ // Is this needed??
+    void set_name(OFString ae_title);
+
+    void setacceptableIPs(OFList<OFString> source_list);
+
+    OFList<OFString> getacceptableIPs();
+
 protected:
     /** Overwrite OFThread's run() method. */
     void run();
+
+    virtual OFCondition listen();
+
 };
